+++ rpc-framework/rpc-core/src/main/java/space/ruiwang/tolerant/impl/Failsafe.java
@@ -3,7 +3,7 @@ package space.ruiwang.tolerant.impl;
 import java.util.List;
 
 import space.ruiwang.domain.RpcRequestConfig;
-import space.ruiwang.domain.RpcRequestDO;
+import space.ruiwang.domain.RpcRequestDTO;
 import space.ruiwang.domain.ServiceMetaData;
 import space.ruiwang.tolerant.FaultTolerant;
 
@@ -13,7 +13,7 @@ import space.ruiwang.tolerant.FaultTolerant;
  */
 public class Failsafe implements FaultTolerant {
     @Override
-    public void handler(RpcRequestDO rpcRequestDO, RpcRequestConfig rpcRequestConfig,
+    public void handler(RpcRequestDTO rpcRequestDTO, RpcRequestConfig rpcRequestConfig,
             List<ServiceMetaData> excludedServices, ServiceMetaData service) {
         rpcRequestConfig.setRetryCount(rpcRequestConfig.getRetryCount() - 1);
     }