+++ rpc-provider2/src/main/java/space/ruiwang/serviceimpl/TestServiceImpl.java
@@ -13,7 +13,7 @@ public class TestServiceImpl implements TestService {
     public String calc(Integer a, Integer b) {
         int c = a + b;
         // 模拟错误
-        int i = 1 / 0;
+//        int i = 1 / 0;
         return "a + b 的和是: " + c;
     }
 }